[Blue Scan] Added Notifications
We now have notifications which appear when we see a new device. Also, I rearranged the code make a lot more sense (like registering the listeners when the app is first created. I have no idea how it was working before). Also, added a simple site that you have to leave open while this is working